fix build

+++ install.sh
@@ -1,13 +1,29 @@
 #!/usr/bin/bash
 
-BASEDIR="$(dirname -- "$(readlink -f -- "$0";)")"
+basedir="$(dirname -- "$(readlink -f -- "$0";)")"
+opt_clean=0
+buildtype="plain"
 
-dest=$BASEDIR/build
-if [[ -d $dest ]]; then
+while (($#)); do
+    case "$1" in
+        clean)
+        opt_clean=1
+        ;;
+        -type)
+        buildtype="debug"
+        ;;
+        *)
+        ;;
+    esac
+    shift
+done
+
+dest=$basedir/build
+if [[ $opt_clean == 1 && -d $dest ]]; then
     rm -rf $dest
 fi
 
-meson setup build -Dbuildtype=plain
+meson setup build -Dbuildtype=${buildtype}
 ninja -C build
 sudo ninja -C build install
